Use swapping rather than looping for randomness
Previously, I was iterating over each slip, and then, in a loop,
generating a random index until I found a box that didn't already have a
loop. In this commit, we switch to an approach wherein we start by
assigning each slip to the box with the same number, and then shuffle
the slips by randomly swapping. This is _much_ faster, and achieves the
same results.